Run baseline on different data

+++ eval_baseline_ncvoter.sh
@@ -5,7 +5,23 @@
 # ###################################################### #
 
 # Baseline
-rm -f ./.engine/.ncvoter*
-EVAL_FLAGS="--gt-labels --no-classifier" ./analyze.sh -n > /its/ksapp002/nohup.log
-EVAL_FLAGS="--gt-labels --no-classifier --full-simvector" ./analyze.sh -n > /its/ksapp002/nohup.log
-EVAL_FLAGS="--gt-labels --no-classifier --parfull-simvector" ./analyze.sh -n > /its/ksapp002/nohup.log
+if [ "$1" == 'trtr' ]; then
+    rm -f ./.engine/.ncvoter*
+    EVAL_FLAGS="--gt-labels --parfull-simvector" ./analyze.sh -n -x train train > /its/ksapp002/nohup.log
+    rm -f ./.engine/.ncvoter*
+    EVAL_FLAGS="--gt-labels --parfull-simvector" ./analyze.sh -n -b train train > /its/ksapp002/nohup.log
+fi
+
+if [ "$1" == 'tete' ]; then
+    rm -f ./.engine/.ncvoter*
+    EVAL_FLAGS="--gt-labels --parfull-simvector" ./analyze.sh -n -x test test > /its/ksapp002/nohup.log
+    rm -f ./.engine/.ncvoter*
+    EVAL_FLAGS="--gt-labels --parfull-simvector" ./analyze.sh -n -b test test > /its/ksapp002/nohup.log
+fi
+
+if [ "$1" == 'teva' ]; then
+    rm -f ./.engine/.ncvoter*
+    EVAL_FLAGS="--gt-labels --parfull-simvector" ./analyze.sh -n -x test test > /its/ksapp002/nohup.log
+    rm -f ./.engine/.ncvoter*
+    EVAL_FLAGS="--gt-labels --parfull-simvector" ./analyze.sh -n -b test test > /its/ksapp002/nohup.log
+fi
